tiktok-profile-scraper/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
selenium==4.18.1
undetected-chromedriver==3.5.5
webdriver-manager==4.0.1
python-dotenv==1.0.1
loguru==0.7.2
requests==2.31.0
certifi==2024.2.2